requirements.txt
Flask==2.2.2
opencensus-ext-azure==1.1.9
gunicorn==20.1.0
